gh-108388: Convert test_concurrent_futures to package (#108401)
Convert test_concurrent_futures to a package of sub-tests.

(cherry picked from commit aa6f787faa4bc45006da4dc2f942fb9b82c98836)
