+++ Lib/test/test_concurrent_futures/test_wait.py
@@ -3,7 +3,6 @@ import threading
 import time
 import unittest
 from concurrent import futures
-from test import support
 
 from .util import (
     CANCELLED_FUTURE, CANCELLED_AND_NOTIFIED_FUTURE, EXCEPTION_FUTURE,
@@ -54,7 +53,6 @@ class WaitTests:
                 finished)
         self.assertEqual(set([future1]), pending)
 
-    @support.requires_resource('walltime')
     def test_first_exception(self):
         future1 = self.executor.submit(mul, 2, 21)
         future2 = self.executor.submit(sleep_and_raise, 1.5)
@@ -112,7 +110,6 @@ class WaitTests:
                               future2]), finished)
         self.assertEqual(set(), pending)
 
-    @support.requires_resource('walltime')
     def test_timeout(self):
         future1 = self.executor.submit(mul, 6, 7)
         future2 = self.executor.submit(time.sleep, 6)
