+++ Misc/NEWS.d/next/Tests/2023-08-24-06-10-36.gh-issue-108388.YCVB0D.rst
@@ -1,0 +1,2 @@
+Convert test_concurrent_futures to a package of 7 sub-tests. Patch by Victor
+Stinner.
